backend/flask/pyproject.toml
[project]
name = "happymapper"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "Flask>=3.0",
  "Flask-Cors>=4.0",
  "firebase-admin>=6.0.0",
  "google-generativeai>=0.3.0",
  "Pillow>=10.0.0",
  "python-dotenv>=1.0",
  "werkzeug>=3.0.0",
] 
